requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
Werkzeug==3.1.7
bcrypt==3.2.2
PyJWT==2.7.0
gunicorn==25.2.0
python-dotenv==1.2.2
alembic==1.18.4
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
